--- a/internal/fs/tree_builder_test.go
+++ b/internal/fs/tree_builder_test.go
@@ -2,7 +2,6 @@ package fs
 
 import (
 	"flag"
-	"fmt"
 	"testing"
 )
 
@@ -11,7 +10,6 @@ var (
 )
 
 func BenchmarkFileTreeBuilder(b *testing.B) {
-	fmt.Printf("path: %s\n", *path)
 	for i := 0; i < b.N; i++ {
 		bl := NewFileTreeBuilder(*path)
 		if err := bl.Build(); err != nil {