## Summary

Use `golang.org/x/sync/errgroup` to perform concurrent directory traversal / listing.  

Fixes #17 

### Impact

Please delete options that are not relevant.

- [x] Performance improvement for file listing

### Testing

**Test Configuration**:
* Go version: `1.18`
* Hardware: `I7 6700HQ, 16Gb Memory DDR4, SSD`
* Operating system: `Linux 5.10.146-1-MANJARO x86_64 GNU/Linux`


### Benchmarks 

```name               old time/op    new time/op    delta
FileTreeBuilder-8     2.65s ± 9%     2.31s ± 4%  -12.95%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
FileTreeBuilder-8     297MB ± 0%     297MB ± 0%     ~     (p=0.280 n=10+10)

name               old allocs/op  new allocs/op  delta
FileTreeBuilder-8     3.00M ± 0%     3.00M ± 0%   -0.00%  (p=0.007 n=10+10)
```