@GaruGaru 2022-10-09
Comments fixed, if we want more performance gains probably a lazy eval for the tree is needed, but I guess that it would require a major refactor to be archived. 

From my benchmarks the vast majority of CPU time is spent on system calls, in particular: `Readdirsnames` and `os.Lstat`.