@ozansz 2022-10-09
> Comments fixed, if we want more performance gains probably a lazy eval for the tree is needed, but I guess that it would require a major refactor to be archived.
> 
> From my benchmarks the vast majority of CPU time is spent on system calls, in particular: `Readdirsnames` and `os.Lstat`.

I agree on the lazy evaluation. It would be a better approach but would definitely require a major update. Thanks for the remarks on `Readdirsnames` and `os.Lstat` too 👍 I will try to find if there’s a way to use less syscalls other than using lazy eval